Rename backend package to avoid Next.js routing conflict
